Update patch set 2
Patch Set 2:

The reason we added the rebuild as a separate flag and not a migrator is that we felt not everyone would require it. We started off with individual migrators for changes in edge rules but later realized that label name changes in Beijing would require us to rebuild all anyways. Given that the probability of this happening in future is bleak, we added utilities into the EdgeMigrator abstract class which would help others upgrading from Casablanca to Dublin to write individual migrators (like for Vertices and Indexes). We explored the option of using the @Enabled annotation to toggle between using and not using this as a migrator, however, given that it needs a rebuild. We thought it would add value to just have a functionality to rebuild all edges (in case someone needs it for whatever the reason is)

Patch-set: 2