Update patch set 2
Patch Set 2:

Ankit, I just noticed you reply to Harish and I now understand the intention of the --rebuildEdges. If I understand it correctly the intention was to have a way to run a "disabled" migration. I think a general functionality like this would benefit the migration controller and would prefer if it was implemented like the -m flag maybe (--runDisabled migrationA migrationB ...) which would add the migrations specified here to be run along with any with the -m.

Patch-set: 2